# prueba-tec/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Estados INEGI",
        "description": "API para sincronizar y consultar estados del INEGI (demo técnica).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor base (APP_URL)"
        }
    ],
    "paths": {
        "/api/states": {
            "get": {
                "tags": [
                    "Estados"
                ],
                "summary": "Listar estados (compatible con DataTables)",
                "description": "Retorna un listado paginado de estados. Acepta paginación por **bloques permitidos** (10,25,50,100) y parámetros de DataTables (`draw`, `start`, `length`, `search[value]`).",
                "operationId": "cc8858fad9a119825d4d20311225ad95",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Búsqueda por nombre/abreviatura/clave",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página (modo API simple)",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "bloque",
                        "in": "query",
                        "description": "Tamaño de página (10,25,50,100). Gana prioridad sobre size/length",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                10,
                                25,
                                50,
                                100
                            ]
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "Tamaño de página (si no se usa 'bloque')",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                10,
                                25,
                                50,
                                100
                            ]
                        }
                    },
                    {
                        "name": "draw",
                        "in": "query",
                        "description": "Flag DataTables",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Offset DataTables",
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "length",
                        "in": "query",
                        "description": "Tamaño DataTables (10,25,50,100)",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                10,
                                25,
                                50,
                                100
                            ]
                        }
                    },
                    {
                        "name": "search[value]",
                        "in": "query",
                        "description": "Texto de búsqueda DataTables",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de estados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/RespuestaListado"
                                        },
                                        {
                                            "$ref": "#/components/schemas/RespuestaDataTables"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/states/{cve_ent}": {
            "get": {
                "tags": [
                    "Estados"
                ],
                "summary": "Detalle de un estado (5 campos)",
                "operationId": "8551c8003e35d61ffdf5f80381cc089b",
                "parameters": [
                    {
                        "name": "cve_ent",
                        "in": "path",
                        "description": "Clave de entidad (01..32)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]{2}$",
                            "example": "19"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del estado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstadoDetalle"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        },
        "/api/states/sync": {
            "post": {
                "tags": [
                    "Estados"
                ],
                "summary": "Sincronizar estados desde INEGI",
                "description": "Descarga y sincroniza los estados (upsert). Si `resynchronize=1` y el INEGI no responde, usa el último snapshot bueno del cache (stale-on-error).",
                "operationId": "b7de82bc1a39838ec95d6ef73d8a0ed4",
                "parameters": [
                    {
                        "name": "resynchronize",
                        "in": "query",
                        "description": "Usar snapshot si INEGI cae (1=true, 0=false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Métricas de sincronización",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SyncRespuesta"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "INEGI no disponible y no hay snapshot"
                    }
                }
            }
        },
        "/api/states/deduplicate": {
            "post": {
                "tags": [
                    "Estados"
                ],
                "summary": "Analizar y eliminar duplicados por cve_ent",
                "description": "Elimina duplicados dejando 1 registro por cve_ent (el de menor id).",
                "operationId": "637abb24e5b23224d9a3c12ec9c017d1",
                "responses": {
                    "200": {
                        "description": "Resultado de la deduplicación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeduplicateRespuesta"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EstadoItem": {
                "required": [
                    "cvegeo",
                    "cve_ent",
                    "nomgeo",
                    "nom_abrev",
                    "pob_total"
                ],
                "properties": {
                    "cvegeo": {
                        "type": "string",
                        "example": "01"
                    },
                    "cve_ent": {
                        "type": "string",
                        "example": "01"
                    },
                    "nomgeo": {
                        "type": "string",
                        "example": "Aguascalientes"
                    },
                    "nom_abrev": {
                        "type": "string",
                        "example": "Ags."
                    },
                    "pob_total": {
                        "type": "integer",
                        "example": 1425607
                    }
                },
                "type": "object"
            },
            "MetaListado": {
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "bloque_aplicado": {
                        "type": "integer",
                        "example": 25
                    },
                    "bloques_permitidos": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            10,
                            25,
                            50,
                            100
                        ]
                    },
                    "total": {
                        "type": "integer",
                        "example": 32
                    },
                    "pages": {
                        "type": "integer",
                        "example": 4
                    }
                },
                "type": "object"
            },
            "RespuestaListado": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/EstadoItem"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/MetaListado"
                    }
                },
                "type": "object"
            },
            "RespuestaDataTables": {
                "properties": {
                    "draw": {
                        "type": "integer",
                        "example": 1
                    },
                    "recordsTotal": {
                        "type": "integer",
                        "example": 32
                    },
                    "recordsFiltered": {
                        "type": "integer",
                        "example": 32
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "cve_ent": {
                                    "type": "string",
                                    "example": "19"
                                },
                                "nomgeo": {
                                    "type": "string",
                                    "example": "Nuevo León"
                                },
                                "nom_abrev": {
                                    "type": "string",
                                    "example": "NL"
                                },
                                "pob_total": {
                                    "type": "integer",
                                    "example": 5784442
                                },
                                "acciones": {
                                    "type": "string",
                                    "example": "http://localhost:8000/api/states/19"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/MetaListado"
                    }
                },
                "type": "object"
            },
            "EstadoDetalle": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/EstadoItem"
                    }
                ]
            },
            "SyncRespuesta": {
                "properties": {
                    "insertados": {
                        "type": "integer",
                        "example": 32
                    },
                    "actualizados": {
                        "type": "integer",
                        "example": 0
                    },
                    "sin_cambios": {
                        "type": "integer",
                        "example": 0
                    },
                    "total_recibidos": {
                        "type": "integer",
                        "example": 32
                    },
                    "resynchronize_param": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "DeduplicateRespuesta": {
                "properties": {
                    "claves_duplicadas": {
                        "type": "integer",
                        "example": 0
                    },
                    "filas_duplicadas": {
                        "type": "integer",
                        "example": 0
                    },
                    "filas_eliminadas": {
                        "type": "integer",
                        "example": 0
                    },
                    "total_final": {
                        "type": "integer",
                        "example": 32
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Estados",
            "description": "Consulta y sincronización de estados"
        }
    ]
}